+++ controllers/dashboard/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const dashboardRoutes = require('./dashboardRoutes');
+
+router.use('/', dashboardRoutes);
+
+module.exports = router;
