+++ controllers/homeRoutes.js
@@ -50,28 +50,28 @@ router.get('/login', (req, res) => {
   res.render('login');
 });
 
-router.get('/dashboard', withAuth, async (req, res) => {
-  try {
-    // Get User's checked owned libraries to display here.
-    const userData = await User.findByPk(req.session.user_id, {
+// router.get('/dashboard', withAuth, async (req, res) => {
+//   try {
+//     // Get User's checked owned libraries to display here.
+//     const userData = await User.findByPk(req.session.user_id, {
 
-      include: [{ model: Library }, { model: LibraryBook }],
-      attributes: { exclude: ['password'] },
-    });
+//       include: [{ model: Library }, { model: LibraryBook }],
+//       attributes: { exclude: ['password'] },
+//     });
 
-    const user = userData.get({ plain: true });
-    console.log(user);
+//     const user = userData.get({ plain: true });
+//     console.log(user);
 
-    res.render('dashboard', {
-      user,
-      logged_in: req.session.logged_in,
-    });
-    console.log('Dashboard Route OK');
-  } catch (err) {
-    res.status(500).json(err);
+//     res.render('dashboard', {
+//       user,
+//       logged_in: req.session.logged_in,
+//     });
+//     console.log('Dashboard Route OK');
+//   } catch (err) {
+//     res.status(500).json(err);
 
-  }
-});
+//   }
+// });
 
 router.get('/libraryinfo/:id', async (req, res) => {
   try {
