+++ controllers/index.js
@@ -1,9 +1,11 @@
 const router = require('express').Router();
+const dashboardRoutes = require('./dashboard');
 
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
+router.use('/dashboard', dashboardRoutes);
 
 module.exports = router;
